--- packages/promptable-query/README.md
@@ -1,16 +1,0 @@
-# GPT React Query
-
-GPT-3 mocking react query
-
-# How to run
-
-make sure you're in gpt-prisma-seed directory. copy .env.example to .env, add your openAI key
-
-all the relevant code is in prisma/seed.ts
-
-run commands:
-
-- pnpm i
-- npx prisma db push
-- npx prisma db seed
-- npx prisma studio
