--- packages/promptable-query/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "promptable-query",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "private": true,
-  "scripts": {
-    "start": "node dist/index.mjs",
-    "dev": "rm -rf dist && tsup ./src/index.ts --format=esm --watch --dts --external react",
-    "lint": "TIMING=1 eslint \"**/*.ts*\"",
-    "lint:ci": "pnpm run lint",
-    "build": "rm -rf dist && tsup ./src/index.ts --format=esm --dts --external react",
-    "build:ci": "pnpm run build"
-  },
-  "keywords": [],
-  "author": "Colin Fortuner",
-  "license": "ISC",
-  "dependencies": {
-    "@promptable/server": "workspace*",
-    "react": "^18.2.0",
-    "react-query": "^3.39.3",
-    "typescript": "latest",
-    "zod": "^3.20.2",
-    "zod-to-ts": "^1.1.2"
-  },
-  "devDependencies": {
-    "tsup": "^6.5.0"
-  }
-}
