--- packages/promptable-query/tsconfig.json
@@ -1,24 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es2017",
-        "lib": [
-            "dom",
-            "dom.iterable",
-            "esnext",
-        ],
-        "module": "esnext",
-        "esModuleInterop": true,
-        "forceConsistentCasingInFileNames": true,
-        "isolatedModules": true,
-        "moduleResolution": "node",
-        "preserveWatchOutput": true,
-        "skipLibCheck": true,
-        "noEmit": true,
-        "strict": true,
-        "baseUrl": ".",
-        "paths": {},
-    },
-    "exclude": [
-        "node_modules"
-    ]
-}
