+++ packages/promptable/core/package.json
@@ -19,7 +19,8 @@
         "jest": "^29.4.1",
         "ts-jest": "^29.0.5",
         "tsup": "^6.5.0",
-        "typescript": "latest"
+        "typescript": "latest",
+        "@types/node-fetch": "^2.6.1"
     },
     "dependencies": {
         "@stdlib/nlp-sentencize": "^0.0.2",
@@ -27,7 +28,7 @@
         "chalk": "^4.1.2",
         "csv-parse": "^5.3.4",
         "gpt3-tokenizer": "^1.1.4",
-        "openai": "^3.1.0",
+        "node-fetch": ">=2.6.7 <3.0.0",
         "uuid": "^9.0.0",
         "zod": "^3.20.2"
     }
