+++ packages/promptable/core/src/chains/MemoryLLMChain.ts
@@ -1,12 +1,8 @@
 import { NoopParser, Parser } from "@prompts/Parser";
 import { Prompt } from "@prompts/Prompt";
 import { CompletionsModelProvider } from "@providers/ModelProvider";
-import { trace } from "../tracing";
 import { Memory } from "src/memories/index";
 import { LLMChain } from "@chains/LLMChain";
-
-// TODO: Trace currently requires an anonymous function to be passed in
-// to the trace function. There is some issue with the binding of this
 
 export class MemoryLLMChain<
   T extends "memory" | "userInput",
@@ -22,17 +18,13 @@ export class MemoryLLMChain<
   }
 
   protected async _run(variables: Record<T, string>) {
-    const formattedPrompt = await trace("prompt.format", (vars) =>
-      this.prompt.format(vars)
-    )(variables);
+    const completion = await this.provider.generate({
+      prompt: this.prompt,
+      variables,
+    });
 
-    const completion = await trace("provider.complete", (p) =>
-      this.provider.generate(p)
-    )(formattedPrompt);
+    const parsed = this.prompt.parse(completion);
 
-    const parsed = await trace("prompt.parse", (c) => this.prompt.parse(c))(
-      completion
-    );
     return parsed;
   }
 
@@ -42,6 +34,6 @@ export class MemoryLLMChain<
       string
     >;
 
-    return await trace("llmchain.run", (v) => this._run(v))(vars);
+    return this._run(vars);
   }
 }
