+++ packages/promptable/server/README.md
@@ -1,67 +1,22 @@
-# Promptable
+# Promptable Server
 
-Promptable is a Typescript library for building fullstack AI applications.
+The Server package provides components to use on your server.
 
-Promptable consists of a collection of utilities and interfaces that help you build applications with LLMs. It is designed to be flexible and extensible so that you can use it with any LLM or Embeddings provider.
+> Note
+> This package is not compatible with browser or edge environments.
 
-The goal of this library is to provide a set of general tools for Typescript and Javascript developers to help them build fullstack AI first applications quickly and easily.
+## Installation
 
-> Right now the library is in early development and is very much experimental. Don't use this in production yet! The API is subject to change as we get feedback.
-> 💻
-
-[Github Repo](https://github.com/cfortuner/promptable).
-[Discord](https://discord.gg/SYmACWTf6V).
-[Twitter](https://twitter.com/promptableai).
-
-## Use Cases:
-
-- 💬 Chatbots & Conversational AI
-- ❓ Question Answering Bots
-- ✍️ Writing Apps
-- 🧑‍✈️ Copilot apps built with Chrome Extensions, VSCode Extensions, and more!
-- 🔍 AI Semantic Search apps
-- 🛠️ AI first Automations, Workflows and Tools
-- 🤖 Autonomous Agents & Personal Assistants
-
-### Features
-
-- [Prompts](./modules/prompts.md) for templating and formatting
-- [Model Providers](./modules/model-providers.md) for Text Generation and Embedding Generation
-- [Embeddings](./modules/embeddings.md) for creating Embeddings, Indexing and Search
-- [Tracing](./modules/tracing.md) for debugging your applications!
-- [Utilities](./modules/utilities.md) for working with text and data.
-
-## Install Library
-
-`npm i promptable`
+```bash
+npm i @promptable/server
+```
 
 ## Usage
 
-See our docs for more info on how to use the library.
-(Documentation)[https://docs-promptable.vercel.app/]
+## Tracing
 
-### Run the Examples
+The tracing package provides a tracing middleware for debugging your server logic
 
-To run an example, clone the repo and run the following commands
-
+```js
+import { tracing } from "@promptable/server";
 ```
-pnpm i
-pnpm dev
-pnpm run example <example name>
-```
-
-### Web UI
-
-To assist in debugging, we also provide a Tracing UI that helps you visualize the steps taken by the LLM to generate the output.
-
-See our (Docs)[https://docs-promptable.vercel.app/docs/modules/tracing#tracing-ui]
-
-## Motivation
-
-Large Language models are emerging as a powerful tool to use for variety of tasks. With OpenAI models like GPT-3 only an API call away, it's become possible to build applications that use AI as a core software component for business logic, data processing, content generation and more. Traditionally, AI tooling has only been built in python to power backend systems, but with the success of ChatGPT, we have learned that the UI/UX of an app is just as important as the backend.
-
-This project aims to provide a set of general tools for Typescript and Javascript developers to help them build fullstack AI first applications.
-
-## Community
-
-If you have any questions about anything related to Promptable or if you want to discuss with us and the community, you are welcome to join our **[discord](https://discord.gg/SYmACWTf6V)**.
