+++ packages/promptable/server/package.json
@@ -22,6 +22,12 @@
         "typescript": "latest"
     },
     "dependencies": {
-        "@promptable/core": "workspace:*"
+        "gpt3-tokenizer": "^1.1.4",
+        "@promptable/core": "workspace:*",
+        "@types/node": "^18.11.18",
+        "axios": "^1.2.4",
+        "chalk": "^4.1.2",
+        "uuid": "^9.0.0",
+        "openai": "latest"
     }
 }
