+++ packages/promptable/server/src/index.ts
@@ -1,1 +1,11 @@
 export {};
+
+// Tracing is only available in nodejs environments for now
+import {
+  graphTraces,
+  trace,
+  setTraceConfig,
+  sendTraceToServer,
+} from "./tracing";
+import type { Trace } from "./tracing";
+export { graphTraces, trace, setTraceConfig, Trace, sendTraceToServer };
